backend/requirements.txt
# backend/requirements.txt
# backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
notebooklm-py[browser]>=0.3.4
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0
ddgs>=9.0.0